Lift the client query override to storage
